chore: move solutions
