+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/[solutionId]/getSolutionIdRouteData.ts
