+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/[solutionId]/page.tsx
