+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/_components/_actions.ts
