+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/_components/solution-detail.tsx
@@ -12,10 +12,10 @@ import { UserBadge } from '@repo/ui/components/user-badge';
 import { Calendar, Flag, Pin, Share, Trash, ArrowLeft, Pencil } from '@repo/ui/icons';
 import clsx from 'clsx';
 import Link from 'next/link';
-import type { ChallengeSolution } from '~/app/[locale]/challenge/[slug]/solutions/[solutionId]/page';
+import type { ChallengeSolution } from '~/app/[locale]/challenge/[slug]/@left/solutions/[solutionId]/page';
 import { ReportDialog } from '~/components/ReportDialog';
 import { getRelativeTime } from '~/utils/relativeTime';
-import { Vote } from '../../../_components/vote';
+import { Vote } from '../../../../_components/vote';
 import { pinOrUnpinSolution } from './_actions';
 import { isAdminOrModerator, isAuthor } from '~/utils/auth-guards';
 import { SolutionDeleteDialog } from './delete';
