+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/_components/solution-skeleton.tsx
