+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/_components/submit-solution.tsx
