+++ apps/web/src/app/[locale]/challenge/[slug]/@left/solutions/getSolutionRouteData.ts
