+++ apps/web/src/app/[locale]/challenge/[slug]/@right/default.tsx
@@ -1,0 +1,3 @@
+import Page from './page';
+
+export default Page;
